Move global handlers out of store-api.
